BVSREC/requirements.txt
altgraph==0.17
anyio==2.0.2
APScheduler==3.6.3
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bitstring==3.1.7
bleach==3.3.0
boto==2.48.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.4
comtypes==1.1.7
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-crispy-forms==1.9.2
django-email-users==1.1.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-appapi==0.0.3
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.3
Faker==6.6.2
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
freegames==2.3.2
future==0.18.2
generics==3.2.1
googletrans==3.0.0
gunicorn==19.9.0
h11==0.9.0
h2==3.2.0
heroku==0.1.4
hpack==3.0.0
hstspreload==2020.10.6
httpcore==0.9.1
httpx==0.13.3
hupper==1.10.2
hyperframe==5.2.0
idna==2.10
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.2.2
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.4
MarkupSafe==1.1.1
mistune==0.8.4
MouseInfo==0.1.3
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.22
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.4
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
PasteDeploy==2.1.1
pefile==2019.4.18
phonenumbers==8.12.20
pickleshare==0.7.5
Pillow==7.2.0
PillowImage==1.1.11
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.1
postgresadapter @ file:///C:/Users/Madhava%20Reddy/Downloads/postgresadapter-2.0.1-cp38-cp38-win32.whl
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
PyAutoGUI==0.9.52
PyBundle==1.0.6
pycparser==2.20
pycryptodome==3.10.1
pygame==1.9.6
PyGetWindow==0.0.9
Pygments==2.7.4
PyInstaller-wheel==3.6
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.1
pypiwin32==223
pyramid==2.0
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.26
pytest==6.2.2
python-dateutil==1.5
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.1
pywhatkit==3.2
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.2
reportlab==3.5.60
requests==2.24.0
rfc3986==1.4.0
Send2Trash==1.5.0
signals==0.0.2
six==1.15.0
sniffio==1.1.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
speedtest-cli==2.1.2
SQLAlchemy==1.3.3
sqlparse==0.3.1
square==0.0.3
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
translationstring==1.4
twilio==6.45.3
tzlocal==2.1
urllib3==1.25.10
users==1.0.dev0
vector==0.0.0
vector-2d==1.5.2
venusian==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==0.15.4
whitenoise==5.2.0
wikipedia==1.4.0
WTForms==2.3.3
zope.deprecation==4.4.0
zope.interface==5.3.0
